Move out ini-load related logic to separate file.
