+++ src/service/resources-impl.js
@@ -306,71 +306,6 @@ export class ResourcesImpl {
     return this.ampdoc;
   }
 
-  /**
-   * Returns a subset of resources which are (1) belong to the specified host
-   * window, and (2) meet the filterFn given.
-   *
-   * TODO(lannka): this method should be moved out of resources service.
-   * @param {!Window} hostWin
-   * @param {function(!Resource):boolean} filterFn
-   * @return {!Promise<!Array<!Resource>>}
-   */
-  getMeasuredResources(hostWin, filterFn) {
-    // First, wait for the `ready-scan` signal. Waiting for each element
-    // individually is too expensive and `ready-scan` will cover most of
-    // the initially parsed elements.
-    // TODO(jridgewell): this path should be switched to use a future
-    // "layer has been measured" signal.
-    return this.ampdoc
-      .signals()
-      .whenSignal(READY_SCAN_SIGNAL_)
-      .then(() => {
-        // Second, wait for any left-over elements to complete measuring.
-        const measurePromiseArray = [];
-        this.resources_.forEach(r => {
-          if (!r.hasBeenMeasured() && r.hostWin == hostWin && !r.hasOwner()) {
-            measurePromiseArray.push(r.getPageLayoutBoxAsync());
-          }
-        });
-        return Promise.all(measurePromiseArray);
-      })
-      .then(() =>
-        this.resources_.filter(r => {
-          return (
-            r.hostWin == hostWin &&
-            !r.hasOwner() &&
-            r.hasBeenMeasured() &&
-            filterFn(r)
-          );
-        })
-      );
-  }
-
-  /**
-   * Returns a subset of resources which are (1) belong to the specified host
-   * window, and (2) positioned in the specified rect.
-   *
-   * TODO(lannka): this method should be moved out of resources service.
-   * @param {!Window} hostWin
-   * @param {!../layout-rect.LayoutRectDef} rect
-   * @param {boolean=} opt_isInPrerender signifies if we are in prerender mode.
-   * @return {!Promise<!Array<!Resource>>}
-   */
-  getResourcesInRect(hostWin, rect, opt_isInPrerender) {
-    return this.getMeasuredResources(hostWin, r => {
-      // TODO(jridgewell): Remove isFixed check here once the position
-      // is calculted correctly in a separate layer for embeds.
-      if (
-        !r.isDisplayed() ||
-        (!r.overlaps(rect) && !r.isFixed()) ||
-        (opt_isInPrerender && !r.prerenderAllowed())
-      ) {
-        return false;
-      }
-      return true;
-    });
-  }
-
   /** @private */
   monitorInput_() {
     const input = Services.inputFor(this.win);
