+++ test/integration/test-amphtml-ads.js
@@ -242,8 +242,7 @@ describe('AMPHTML ad on non-AMP page (inabox)', () => {
   );
 });
 
-// TODO(wg-ads, #24421): Flaky in experiment B (INABOX_LITE) on Chrome.
-describe.skip('A more real AMPHTML image ad', () => {
+describe('A more real AMPHTML image ad', () => {
   const {testServerPort} = window.ampTestRuntimeConfig;
 
   // The image ad as seen in examples/inabox.gpt.html,
